rename loginfilter to gateway
